+++ backend/app/Migration.hs
@@ -3,24 +3,25 @@ module Main where
 import Amazonka qualified as AWS
 import AppConfig (AppM, Env (..), authDatabaseFile, databaseFile, redisConfig, runAppLogger)
 import AppServer (insertReport_, reprocessReports)
-import Control.Monad.Logger (LogLevel (..))
+import Control.Monad.Logger (LogLevel (..), MonadLogger)
 import Data.Csv (FromRecord, HasHeader (..), decode)
 import Data.Validation (Validation (..))
 import Data.Vector qualified as V
-import Database (insertLegacyEntry, runDb)
-import Database.Esqueleto.Experimental (defaultConnectionPoolConfig)
+import Database (DBAction, insertInitialStats, insertPlayerIfNotExists, repsertPlayerStats, runDb)
+import Database.Esqueleto.Experimental (Entity (..), defaultConnectionPoolConfig)
 import Database.Persist.Sqlite (createSqlitePoolWithConfig)
 import Database.Redis (connect)
-import Logging (Logger, log, stdoutLogger, (<>:))
+import Logging (Logger, stdoutLogger, (<>:))
+import Logging qualified as L
 import Servant (ServerError (errBody), err500, runHandler, throwError)
 import System.Directory (createDirectoryIfMissing)
 import System.Environment (setEnv)
 import System.FilePath (takeDirectory)
 import Types.Api (RawGameReport (victory))
 import Types.DataField (Victory (..))
-import Types.Database (migrateSchema)
+import Types.Database (PlayerStatsInitial (..), PlayerStatsTotal (..), defaultPlayerStatsYear, migrateSchema)
 import Types.Migration
-  ( ParsedGameReport (timestamp),
+  ( ParsedGameReport (log, timestamp),
     ParsedLegacyLadderEntry (..),
     PlayerBanList,
     toParsedGameReport,
@@ -32,16 +33,27 @@ import Validation (ReportError (..), validateReport)
 banList :: PlayerBanList
 banList = ["mordak", "mellowsedge"]
 
+insertLegacyEntry :: (MonadIO m, MonadLogger m) => ParsedLegacyLadderEntry -> DBAction m ()
+insertLegacyEntry entry = do
+  (Entity playerId _) <- insertPlayerIfNotExists entry.player
+
+  let initialStats = PlayerStatsInitial playerId entry.freeRating entry.shadowRating entry.gamesPlayedTotal
+  let totalStats = PlayerStatsTotal playerId entry.freeRating entry.shadowRating entry.gamesPlayedTotal
+  let yearStats = defaultPlayerStatsYear playerId 2022
+
+  insertInitialStats initialStats
+  repsertPlayerStats (totalStats, yearStats)
+
 migrate :: [ParsedLegacyLadderEntry] -> [ParsedGameReport] -> AppM ()
 migrate legacyEntries reports = runDb $ do
   traverse_ insertLegacyEntry . filter (\entry -> entry.player `notElem` banList) $ legacyEntries
 
-  forM_ (map (\r -> (r.timestamp, toRawGameReport r)) reports) $ \(timestamp, report) -> do
+  forM_ (map (\r -> (r.timestamp, toRawGameReport r, r.log)) reports) $ \(timestamp, report, s3Url) -> do
     case validateReport report of
       Failure errs -> case errs of
-        [NoVictoryConditionMet] -> insertReport_ timestamp (report {victory = Concession}) Nothing
+        [NoVictoryConditionMet] -> insertReport_ timestamp (report {victory = Concession}) s3Url
         _ -> throwError $ err500 {errBody = "Unrecognized failure: " <>: errs <> " for report: " <>: report}
-      Success _ -> insertReport_ timestamp report Nothing
+      Success _ -> insertReport_ timestamp report s3Url
 
   reprocessReports
 
@@ -55,7 +67,7 @@ tryParse path logger mapper = do
 
   case parsed of
     Left err -> do
-      log logger LevelError $ "Error parsing " <>: path <> ": " <>: err
+      L.log logger LevelError $ "Error parsing " <>: path <> ": " <>: err
       pure Nothing
     Right a -> pure $ Just a
 
@@ -74,13 +86,13 @@ main = do
   let env = Env {dbPool, authDbPool, redisPool, logger, aws}
 
   legacyEntries <- tryParse "migration/legacy-ladder.csv" logger toParsedLegacyLadderEntry
-  reports <- tryParse "migration/reports.csv" logger toParsedGameReport
+  reports <- tryParse "migration/reports.csv" logger (toParsedGameReport aws)
 
   case (legacyEntries, reports) of
     (Just es, Just rs) -> do
       migrateSchema dbPool logger
       migrationResult <- runHandler . runAppLogger logger . usingReaderT env $ migrate es (reverse rs)
       case migrationResult of
-        Left err -> log logger LevelError $ "Migration failed: " <>: err
+        Left err -> L.log logger LevelError $ "Migration failed: " <>: err
         Right _ -> pass
     _ -> pass
